bugfix updateAdapt

+++ lib/Class_Para_Tree.cpp
@@ -339,7 +339,7 @@ void Class_Para_Tree::updateAdapt() {
 		uint64_t local_num_octants = octree.getNumOctants();
 		error_flag = MPI_Allreduce(&local_num_octants,&global_num_octants,1,MPI_UINT64_T,MPI_SUM,MPI_COMM_WORLD);
 		//update partition_range_globalidx
-		int rbuff [nproc];
+		uint64_t rbuff [nproc];
 		error_flag = MPI_Allgather(&local_num_octants,1,MPI_UINT64_T,&rbuff,1,MPI_UINT64_T,MPI_COMM_WORLD);
 		for(int p = 0; p < nproc; ++p){
 			partition_range_globalidx[p] = 0;
